chore(toolbar): cut the preflight and web-analytics-achievements edges
Continues .agents/toolbar-migration.md (Step 6):

- preflightLogic moves from scenes/PreflightCheck/ to lib/logic/preflightLogic - it is global app state (instance status/realm), not scene state, and five lib files shipped in the toolbar bundle depend on it. All src/lib importers are repointed; a re-export at the old path keeps the ~100 scene/product importers working, so the full repoint can happen as later cleanup. Cuts 5 baseline edges.
- eventUsageLogic no longer reaches into web-analytics achievements: webAnalyticsAchievementsLogic now owns interaction recording via a recordInteraction action (absorbing recordInteraction.ts, deleted) and listens to the five reportWebAnalytics* action types itself. The feature owns its triggers, and the toolbar graph drops the achievements modules plus the generated web_analytics API client. Cuts 2 baseline edges.

Graph baseline 25 -> 18 edges; 8 files (~62 KB source) leave the toolbar graph.
