+++ .agents/toolbar-migration.md
@@ -4,7 +4,7 @@ This file is the committed source of truth for the toolbar bundle modernization 
 
 ## Status
 
-- **Current step**: Steps 1–4 and 7 implemented and in review as a Graphite stack (drafts), plus a seventh PR batching the simple Step 5/6 moves (PanelSettings, UrlMatchingHints, KeyboardShortcut, LemonMarkdown mermaid barrel — baseline 31 → 25 edges, mermaid deny retired). Next: merge the base of the stack before extending further, then the remaining Step 6 seams (Link→DraggableToNotebook, LemonColor, eventUsageLogic) and Step 8.
+- **Current step**: Steps 1–4 and 7 implemented and in review as a Graphite stack (drafts), plus a seventh PR batching the simple Step 5/6 moves (PanelSettings, UrlMatchingHints, KeyboardShortcut, LemonMarkdown mermaid barrel — baseline 31 → 25) and an eighth cutting the preflight + web-analytics-achievements edges (baseline 25 → 18). Next: merge the base of the stack before extending further, then the remaining Step 6 seams (Link→DraggableToNotebook, LemonColor→theme logics, taxonomy/PropertyKeyInfo→surveyQuestionLabelsLogic) and Step 8.
 - **Last updated**: 2026-07-07
 - **PRs**: Step 1 — `rafa/toolbar-migration-1-graph-guard` (#69045); Step 3 — `rafa/toolbar-migration-2-products-urls-split` (#69071); chain cuts + shim-leak fix — `rafa/toolbar-migration-3-shim-leak-and-chain-cuts` (#69088); Step 2 — `rafa/toolbar-migration-4-loader-esm-splitting` (#69093); Step 4 — `rafa/toolbar-migration-5-types-replay-shared-cut` (#69106); Step 7 — `rafa/toolbar-migration-6-lazy-menus` (#69113); Step 5/6 batch — `rafa/toolbar-migration-7-simple-edge-cuts` (#69122) (each stacked on the previous)
 - **Reorder note (resolved)**: a trial Step 2 build emitted **487 dead chunks** — with `products.tsx` reachable, every product scene's dynamic import becomes a real chunk in `dist/toolbar/`. The gate was: flip the format only once `products.tsx` leaves the graph. That's now done via two moves in the third PR. First, the remaining chains all funneled through one root cause: the kea shims only matched exact alias strings (`scenes/teamLogic`), so relative imports (`dataThemeLogic` → `./teamLogic`) and `~/`-prefixed imports pulled the REAL logics and their whole app graph — fixed with relative + `~/` matching in `createToolbarModulePlugin`, disconnecting `products.tsx`, `scenes.ts`, `teamLogic`, `PersonDisplay`, and `experimentsLogic` in one move. Second, `scenes/urls` itself (imported by lib components shared with the app) is shimmed to the toolbar's parity-tested `~/toolbar/urls` duplicate from Step 3, taking the last products-manifest path out. Bundle graph 8448 → ~1000 files (99 → ~13.5 MiB source), shipped `dist/toolbar.js` 9.95 MB → ~4 MB (was just under the 10 MB CloudFront cliff). Graph budget ratcheted 110 MB → 18 MB.
@@ -110,7 +110,9 @@ Approach (revised on review): rather than restructuring the generated products m
 - [x] `lib/lemon-ui/LemonMarkdown/index.ts`: the barrel no longer re-exports `LemonMarkdownWithMermaid` (the one product user imports the module directly); the `mermaid` deny in `toolbar-config.mjs` is retired, reintroduction caught by FORBIDDEN_PACKAGES
 - [x] `HeatmapEventsPanel` → `PersonDisplay`: replaced with a plain span (PersonDisplay renders exactly that for a distinct_id-only person with `noPopover`); `TZLabel` → teamLogic edge now shimmed at resolve time (`~/scenes/teamLogic` matching)
 - [x] `KeyboardShortcut` moved out of `layout/navigation-3000/components/` into `lib/components/KeyboardShortcut/` (25 importers repointed; cuts the 4 lemon-ui → layout edges)
-- [ ] `TZLabel` → urls (already shim-resolved to `~/toolbar/urls`; boundary-only); `eventUsageLogic` → preflight/web-analytics/product-tours edges
+- [x] `preflightLogic` moved from `scenes/PreflightCheck/` to `lib/logic/preflightLogic` (it's global app state, not scene state; its `scenes/urls` import resolves through the toolbar shim). All `src/lib/**` importers repointed; a `scenes/PreflightCheck/preflightLogic.tsx` re-export keeps the other ~100 app importers working, so a follow-up repoint is optional cleanup. Cuts 5 baseline edges (incidentStatusLogic, MCPInstallCommand, payGateMiniLogic, superpowersLogic, eventUsageLogic)
+- [x] `eventUsageLogic` → web-analytics achievements: inverted — `webAnalyticsAchievementsLogic` now owns interaction recording (a `recordInteraction` action absorbing `recordInteraction.ts`, deleted) and listens to the five `reportWebAnalytics*` action types itself. Cuts 2 edges and drops the achievements + generated web_analytics API modules (~62 KB, 8 files) from the toolbar graph. Behavior note: recording now requires the achievements logic mounted, which the web analytics scene menu bar guarantees wherever those reports fire
+- [ ] `TZLabel` → urls (already shim-resolved to `~/toolbar/urls`; boundary-only); `eventUsageLogic` → `product-tours/constants` (1 KB leaf, part of the product-tours seam family)
 
 ### Step 7 — lazy boundaries in toolbar source (in review: `rafa/toolbar-migration-6-lazy-menus`)
