+++ frontend/bin/toolbar-graph-baseline.json
@@ -2,21 +2,14 @@
     "//": "Toolbar -> app-zone import edges that exist today, enforced by bin/check-toolbar-graph.mjs. Do not add entries; delete them as the migration (.agents/toolbar-migration.md) cuts each edge.",
     "edges": [
         "src/lib/components/HedgehogMode/HedgehogMode.tsx -> src/layout/navigation-3000/themeLogic.ts",
-        "src/lib/components/HelpMenu/incidentStatusLogic.tsx -> src/scenes/PreflightCheck/preflightLogic.tsx",
         "src/lib/components/JSONViewer.tsx -> src/layout/navigation-3000/themeLogic.ts",
-        "src/lib/components/MCPHint/MCPInstallCommand.tsx -> src/scenes/PreflightCheck/preflightLogic.tsx",
-        "src/lib/components/PayGateMini/payGateMiniLogic.tsx -> src/scenes/PreflightCheck/preflightLogic.tsx",
         "src/lib/components/PropertyKeyInfo.tsx -> src/scenes/surveys/surveyQuestionLabelsLogic.ts",
-        "src/lib/components/Superpowers/superpowersLogic.ts -> src/scenes/PreflightCheck/preflightLogic.tsx",
         "src/lib/lemon-ui/LemonColor/LemonColorButton.tsx -> src/scenes/dataThemeLogic.tsx",
         "src/lib/lemon-ui/LemonColor/LemonColorGlyph.tsx -> src/layout/navigation-3000/themeLogic.ts",
         "src/lib/lemon-ui/LemonColor/LemonColorGlyph.tsx -> src/scenes/dataThemeLogic.tsx",
         "src/lib/lemon-ui/LemonColor/LemonColorPicker.tsx -> src/scenes/dataThemeLogic.tsx",
         "src/lib/lemon-ui/Link/Link.tsx -> src/scenes/notebooks/AddToNotebook/DraggableToNotebook.tsx",
-        "src/lib/utils/eventUsageLogic.ts -> ../products/web_analytics/frontend/generated/api.schemas.ts",
-        "src/lib/utils/eventUsageLogic.ts -> src/scenes/PreflightCheck/preflightLogic.tsx",
         "src/lib/utils/eventUsageLogic.ts -> src/scenes/product-tours/constants.ts",
-        "src/lib/utils/eventUsageLogic.ts -> src/scenes/web-analytics/achievements/recordInteraction.ts",
         "src/taxonomy/helpers.ts -> src/scenes/surveys/surveyQuestionLabelsLogic.ts",
         "src/taxonomy/taxonomy.tsx -> src/scenes/web-analytics/tabs/marketing-analytics/utils.ts",
         "src/toolbar/experiments/experimentsTabLogic.tsx -> src/scenes/experiments/experimentStatus.ts",
