+++ frontend/src/lib/components/EventPropertyTabs/eventPropertyFilteringLogic.ts
@@ -1,7 +1,7 @@
 import { connect, kea, path, selectors } from 'kea'
 
+import { preflightLogic } from 'lib/logic/preflightLogic'
 import { userPreferencesLogic } from 'lib/logic/userPreferencesLogic'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 
 import { CLOUD_INTERNAL_POSTHOG_PROPERTY_KEYS, PROPERTY_KEYS } from '~/taxonomy/taxonomy'
