+++ frontend/src/lib/components/HelpMenu/incidentStatusLogic.tsx
@@ -4,7 +4,7 @@ import posthog from 'posthog-js'
 
 // eslint-disable-next-line import/no-cycle
 import { superpowersLogic } from 'lib/components/Superpowers/superpowersLogic'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
+import { preflightLogic } from 'lib/logic/preflightLogic'
 
 import { Region } from '~/types'
