+++ frontend/src/lib/components/MarkdownEditor/inline/inlineMarkdownSlashCommands.test.tsx
@@ -4,7 +4,7 @@ import { act, cleanup, fireEvent, render, screen } from '@testing-library/react'
 import type { Editor } from '@tiptap/core'
 import { createRef, type ComponentRef } from 'react'
 
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
+import { preflightLogic } from 'lib/logic/preflightLogic'
 
 import { initKeaTests } from '~/test/init'
 import { Realm, type PreflightStatus } from '~/types'
