+++ frontend/src/lib/components/MarkdownEditor/shared/MarkdownEditorImageEmojiControls.tsx
@@ -10,7 +10,7 @@ import { LemonFileInput } from 'lib/lemon-ui/LemonFileInput'
 import { emojiUsageLogic } from 'lib/lemon-ui/LemonTextArea/emojiUsageLogic'
 import { Spinner } from 'lib/lemon-ui/Spinner'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
+import { preflightLogic } from 'lib/logic/preflightLogic'
 
 import { useMarkdownEditorImageUpload } from './useMarkdownEditorImageUpload'
