+++ frontend/src/lib/components/NavPanelAdvertisement/NavPanelAdvertisement.tsx
@@ -1,7 +1,7 @@
 import { useValues } from 'kea'
 
 import { getFeatureFlagPayload } from 'lib/logic/featureFlagLogic'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
+import { preflightLogic } from 'lib/logic/preflightLogic'
 import { userLogic } from 'scenes/userLogic'
 
 import { panelLayoutLogic } from '~/layout/panel-layout/panelLayoutLogic'
