+++ frontend/src/lib/components/Superpowers/Superpowers.tsx
@@ -5,9 +5,9 @@ import { LemonButton, LemonDivider, LemonSelect, LemonSwitch, LemonTag } from '@
 
 import { SupermanHog } from 'lib/components/hedgehogs'
 import { LemonModal } from 'lib/lemon-ui/LemonModal'
+import { preflightLogic } from 'lib/logic/preflightLogic'
 import { maxGlobalLogic } from 'scenes/max/maxGlobalLogic'
 import { organizationLogic } from 'scenes/organizationLogic'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 import { teamLogic } from 'scenes/teamLogic'
 import { userLogic } from 'scenes/userLogic'
