+++ frontend/src/lib/components/UpgradeModal/upgradeModalLogic.ts
@@ -1,8 +1,8 @@
 import { actions, connect, kea, listeners, path, reducers, selectors } from 'kea'
 
+import { preflightLogic } from 'lib/logic/preflightLogic'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { billingLogic } from 'scenes/billing/billingLogic'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 import { userLogic } from 'scenes/userLogic'
 
 import { AvailableFeature } from '~/types'
