+++ frontend/src/lib/integrations/slackIntegrationLogic.ts
@@ -4,7 +4,7 @@ import { loaders } from 'kea-loaders'
 import api from 'lib/api'
 import { dayjs } from 'lib/dayjs'
 import { getRecentSlackChannelIds, slackChannelId } from 'lib/integrations/slackChannel'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
+import { preflightLogic } from 'lib/logic/preflightLogic'
 
 import { SlackChannelType } from '~/types'
