+++ frontend/src/lib/lemon-ui/LemonTextArea/LemonTextAreaMarkdown.tsx
@@ -15,7 +15,7 @@ import { LemonTextArea, LemonTextAreaProps } from 'lib/lemon-ui/LemonTextArea/Le
 import { lemonToast } from 'lib/lemon-ui/LemonToast'
 import { Spinner } from 'lib/lemon-ui/Spinner'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
+import { preflightLogic } from 'lib/logic/preflightLogic'
 
 export const LemonTextAreaMarkdown = React.forwardRef<HTMLTextAreaElement, LemonTextAreaProps>(
     function LemonTextAreaMarkdown({ value, onChange, className, ...editAreaProps }, ref): JSX.Element {
