+++ frontend/src/scenes/web-analytics/CrossSellButtons/ReplayButton.tsx
@@ -2,7 +2,7 @@ import posthog from 'posthog-js'
 
 import ViewRecordingsPlaylistButton from 'lib/components/ViewRecordingButton/ViewRecordingsPlaylistButton'
 import { addProductIntentForCrossSell } from 'lib/utils/product-intents'
-import { recordWebAnalyticsInteraction } from 'scenes/web-analytics/achievements/recordInteraction'
+import { webAnalyticsAchievementsLogic } from 'scenes/web-analytics/achievements/webAnalyticsAchievementsLogic'
 import { BREAKDOWN_NULL_DISPLAY } from 'scenes/web-analytics/common'
 
 import {
@@ -146,7 +146,7 @@ export const ReplayButton = ({
     const handleClick = (e: React.MouseEvent): void => {
         e.stopPropagation()
         posthog.capture('web_analytics_recording_opened', { breakdown_by: breakdownBy })
-        recordWebAnalyticsInteraction(InteractionKindEnumApi.Recording)
+        webAnalyticsAchievementsLogic.findMounted()?.actions.recordInteraction(InteractionKindEnumApi.Recording)
         void addProductIntentForCrossSell({
             from: ProductKey.WEB_ANALYTICS,
             to: ProductKey.SESSION_REPLAY,
