+++ frontend/src/scenes/web-analytics/achievements/webAnalyticsAchievementsLogic.test.ts
@@ -1,6 +1,9 @@
 import { expectLogic } from 'kea-test-utils'
 
+import { FEATURE_FLAGS } from 'lib/constants'
 import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
+import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 
 import { useMocks } from '~/mocks/jest'
 import { initKeaTests } from '~/test/init'
@@ -12,6 +15,7 @@ import { webAnalyticsAchievementsLogic } from './webAnalyticsAchievementsLogic'
 const OVERVIEW_URL = '/api/projects/:team_id/web_analytics_achievements/overview/'
 const ACKNOWLEDGE_URL = '/api/projects/:team_id/web_analytics_achievements/acknowledge_celebration/'
 const PREFERENCES_URL = '/api/projects/:team_id/web_analytics_achievements/preferences/'
+const RECORD_INTERACTION_URL = '/api/projects/:team_id/web_analytics_achievements/record_interaction/'
 
 const MOCK_OVERVIEW: AchievementsListResponseApi = {
     definitions: [],
@@ -59,9 +63,11 @@ const DEFS_OVERVIEW: AchievementsListResponseApi = {
 describe('webAnalyticsAchievementsLogic', () => {
     let logic: ReturnType<typeof webAnalyticsAchievementsLogic.build>
     let lastAck: { track_key: string; stage: number } | null
+    let lastInteraction: { interaction_kind: string } | null
 
     beforeEach(() => {
         lastAck = null
+        lastInteraction = null
         jest.spyOn(lemonToast, 'success').mockReturnValue('mock-toast-id')
         initKeaTests()
         useMocks({
@@ -73,6 +79,10 @@ describe('webAnalyticsAchievementsLogic', () => {
                 [ACKNOWLEDGE_URL]: async ({ request }) => {
                     lastAck = (await request.json()) as { track_key: string; stage: number }
                     return [200, { acknowledged: true }]
+                },
+                [RECORD_INTERACTION_URL]: async ({ request }) => {
+                    lastInteraction = (await request.json()) as { interaction_kind: string }
+                    return [200, { recorded: true }]
                 },
             },
         })
@@ -202,6 +212,45 @@ describe('webAnalyticsAchievementsLogic', () => {
 
         expect(logic.values.sortedUserTracks.map((track) => track.key)).toEqual(['explorer', 'detective', 'streak'])
         expect(logic.values.sortedTeamTracks.map((track) => track.key)).toEqual(['traffic'])
+    })
+
+    describe('interaction recording', () => {
+        const enableAchievements = (): void => {
+            featureFlagLogic.actions.setFeatureFlags([], {
+                [FEATURE_FLAGS.WEB_ANALYTICS_ACHIEVEMENTS]: true,
+                [FEATURE_FLAGS.WEB_ANALYTICS_STREAK_CADENCE]: 'daily',
+            })
+        }
+
+        it.each([
+            'reportWebAnalyticsFilterApplied',
+            'reportWebAnalyticsFilterRemoved',
+            'reportWebAnalyticsDateRangeChanged',
+            'reportWebAnalyticsCompareToggled',
+            'reportWebAnalyticsPathCleaningToggled',
+        ] as const)('records a data interaction when %s fires', async (report) => {
+            enableAchievements()
+            await expectLogic(logic, () => {
+                eventUsageLogic.mount()
+                eventUsageLogic.actions[report]({} as any)
+            })
+                .toDispatchActions(['recordInteraction'])
+                .toFinishAllListeners()
+            expect(lastInteraction).toEqual({ interaction_kind: 'data' })
+        })
+
+        it('does not post an interaction when achievements are disabled', async () => {
+            featureFlagLogic.actions.setFeatureFlags([], {
+                [FEATURE_FLAGS.WEB_ANALYTICS_ACHIEVEMENTS]: false,
+            })
+            await expectLogic(logic, () => {
+                eventUsageLogic.mount()
+                eventUsageLogic.actions.reportWebAnalyticsFilterApplied({} as any)
+            })
+                .toDispatchActions(['recordInteraction'])
+                .toFinishAllListeners()
+            expect(lastInteraction).toBeNull()
+        })
     })
 
     it('toggles expanded tracks and clears them when the modal closes', async () => {
