+++ frontend/src/scenes/web-analytics/achievements/webAnalyticsAchievementsLogic.ts
@@ -4,17 +4,20 @@ import posthog from 'posthog-js'
 
 import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { teamLogic } from 'scenes/teamLogic'
 
 import {
     webAnalyticsAchievementsAcknowledgeCelebration,
     webAnalyticsAchievementsOverview,
+    webAnalyticsAchievementsRecordInteraction,
 } from 'products/web_analytics/frontend/generated/api'
-import type {
-    AchievementDefinitionApi,
-    AchievementProgressApi,
-    AchievementsListResponseApi,
-    PendingCelebrationApi,
+import {
+    type AchievementDefinitionApi,
+    type AchievementProgressApi,
+    type AchievementsListResponseApi,
+    InteractionKindEnumApi,
+    type PendingCelebrationApi,
 } from 'products/web_analytics/frontend/generated/api.schemas'
 
 import { deriveTrackProgress } from './achievementProgress'
@@ -55,6 +58,7 @@ export const webAnalyticsAchievementsLogic = kea<webAnalyticsAchievementsLogicTy
         openModal: true,
         closeModal: true,
         acknowledgeCelebration: (trackKey: string, stage: number) => ({ trackKey, stage }),
+        recordInteraction: (kind: InteractionKindEnumApi) => ({ kind }),
         markCelebrated: (key: string) => ({ key }),
         triggerConfetti: true,
         toggleTrackExpanded: (trackKey: string) => ({ trackKey }),
@@ -217,6 +221,39 @@ export const webAnalyticsAchievementsLogic = kea<webAnalyticsAchievementsLogicTy
             if (isWebAnalyticsAchievementsEnabled(values.featureFlags, values.achievementsOptOut)) {
                 actions.loadAchievements()
             }
+        },
+        recordInteraction: async ({ kind }) => {
+            if (!isWebAnalyticsAchievementsEnabled(values.featureFlags, values.achievementsOptOut)) {
+                return
+            }
+            if (values.currentProjectId === undefined || values.currentProjectId === null) {
+                return
+            }
+            try {
+                await webAnalyticsAchievementsRecordInteraction(String(values.currentProjectId), {
+                    interaction_kind: kind,
+                })
+                actions.loadAchievements()
+            } catch {
+                // best-effort gamification signal: never surface recording failures
+            }
+        },
+        // Dashboard interactions count toward achievements. These reports fire only from web
+        // analytics UI, where this logic is mounted via the scene menu bar.
+        [eventUsageLogic.actionTypes.reportWebAnalyticsFilterApplied]: () => {
+            actions.recordInteraction(InteractionKindEnumApi.Data)
+        },
+        [eventUsageLogic.actionTypes.reportWebAnalyticsFilterRemoved]: () => {
+            actions.recordInteraction(InteractionKindEnumApi.Data)
+        },
+        [eventUsageLogic.actionTypes.reportWebAnalyticsDateRangeChanged]: () => {
+            actions.recordInteraction(InteractionKindEnumApi.Data)
+        },
+        [eventUsageLogic.actionTypes.reportWebAnalyticsCompareToggled]: () => {
+            actions.recordInteraction(InteractionKindEnumApi.Data)
+        },
+        [eventUsageLogic.actionTypes.reportWebAnalyticsPathCleaningToggled]: () => {
+            actions.recordInteraction(InteractionKindEnumApi.Data)
         },
     })),
     afterMount(({ actions, values }) => {
